requirements.txt
aspy.yaml==1.3.0
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
distlib==0.3.5
executing==0.9.1
filelock==3.7.1
Flask==2.1.0
flask-googlemaps
Flask-Mobility==1.1.0
ghp-import==2.1.0
htmlmin==0.1.12
identify==2.5.2
idna==3.3
importlib-metadata==4.12.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsmin==3.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mergedeep==1.3.4
mkdocs==1.3.1
mkdocs-material==5.5.14
mkdocs-material-extensions==1.0.3
mkdocs-minify-plugin==0.3.0
nodeenv==1.7.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.2
pre-commit==1.21.0
prompt-toolkit==3.0.30
pure-eval==0.2.2
Pygments==2.12.0
pymdown-extensions==9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
pyyaml-env-tag==0.1
requests==2.28.1
six==1.16.0
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
urllib3==1.26.11
virtualenv==20.16.2
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.1
zipp==3.8.1
